+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Termplot Changelog
 
+- Remove commmand line `--width` and `--height` options (must use plotly json)
 - Use plotly `width` and `height` for browser size